requirements.txt
pyairtable>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0  # For better performance with fuzzywuzzy
python-dateutil>=2.8.2
python-dotenv>=1.0.0  # For loading environment variables from .env file